Struct amethyst_ui::UiButton

pub struct UiButton { /* fields omitted */ }

A clickable button, this must be paired with a UiImage and this entity must have a child entity with a UiText.

Methods

impl UiButton

pub fn new(
normal_text_color: [f32; 4],
hover_text_color: Option<[f32; 4]>,
press_text_color: Option<[f32; 4]>
) -> Self

A constructor for this component. It's recommended to use either a prefab or UiButtonBuilder rather than this function if possible.