Struct amethyst_ui::UiText

pub struct UiText {
    pub text: String,
    pub font_size: f32,
    pub color: [f32; 4],
    pub font: FontHandle,
    pub password: bool,
    // some fields omitted
}

A component used to display text in this entity's UiTransform

Fields
text: String

The string rendered by this.

font_size: f32

The height of a line of text in pixels.

color: [f32; 4]

The color of the rendered text, using a range of 0.0 to 1.0 per channel.

font: FontHandle

The font used for rendering.

password: bool

If true this will be rendered as dots instead of the text.

Methods

impl UiText

pub fn new(
font: FontHandle,
text: String,
color: [f32; 4],
font_size: f32
) -> UiText

Initializes a new UiText

Parameters

font: A handle to a Font asset

text: the glyphs to render

color: RGBA color with a maximum of 1.0 and a minimum of 0.0 for each channel

font_size: a uniform scale applied to the glyphs