Struct amethyst_ui::UiTransform

pub struct UiTransform {
    pub id: String,
    pub anchor: Anchor,
    pub stretch: Stretch,
    pub local_x: f32,
    pub local_y: f32,
    pub local_z: f32,
    pub width: f32,
    pub height: f32,
    pub tab_order: i32,
    pub scale_mode: ScaleMode,
    pub opaque: bool,
    // some fields omitted
}

The UiTransform represents the transformation of a ui element. Values are in pixel and the position is calculated from the bottom left of the screen to the center of the ui element's area.

Fields
id: String

An identifier. Serves no purpose other than to help you distinguish between UI elements.

anchor: Anchor

Indicates where the element sits, relative to the parent (or to the screen, if there is no parent)

stretch: Stretch

If a child ui element needs to fill its parent this can be used to stretch it to the appropriate size.

local_x: f32

X coordinate, 0 is the left edge of the screen. If scale_mode is set to pixel then the width of the screen in pixel is the right edge. If scale_mode is percent then the right edge is 1.

Centered in the middle of the ui element.

local_y: f32

Y coordinate, 0 is the bottom edge of the screen. If scale_mode is set to pixel then the height of the screen in pixel is the top edge. If scale_mode is percent then the top edge is 1.

Centered in the middle of the ui element.

local_z: f32

Z order, entities with a higher Z order will be rendered on top of entities with a lower Z order.

width: f32

The width of this UI element.

height: f32

The height of this UI element.

tab_order: i32

The UI element tab order. When the player presses tab the UI focus will shift to the UI element with the next highest tab order, or if another element with the same tab_order as this one exists they are ordered according to Entity creation order. Shift-tab walks this ordering backwards.

scale_mode: ScaleMode

The scale mode indicates if the position is in pixel or is relative (%) (WIP!) to the parent's size.

opaque: bool

Indicates if actions on the ui can go through this element. If set to false, the element will behaves as if it was transparent and will let events go to the next element (for example, the text on a button).

Methods

impl UiTransform

pub fn new(
id: String,
anchor: Anchor,
x: f32,
y: f32,
z: f32,
width: f32,
height: f32,
tab_order: i32
) -> UiTransform

Creates a new UiTransform. By default, it is considered opaque.

pub fn position_inside_local(&self, x: f32, y: f32) -> bool

Checks if the input position is in the UiTransform rectangle. Uses local coordinates (ignores layouting).

pub fn position_inside(&self, x: f32, y: f32) -> bool

Checks if the input position is in the UiTransform rectangle.

pub fn as_percent(self) -> Self

Renders this UI element by evaluating transform as a percentage of the parent size, rather than rendering it with pixel units.

pub fn as_transparent(self) -> Self

Sets the opaque variable to false, allowing ui events to go through this ui element.

pub fn with_stretch(self, stretch: Stretch) -> Self

Adds stretching to this ui element so it can fill its parent.

pub fn pixel_x(&self) -> f32

Returns the global x coordinate of this UiTransform as computed by the UiTransformSystem.

pub fn pixel_y(&self) -> f32

Returns the global y coordinate of this UiTransform as computed by the UiTransformSystem.

pub fn global_z(&self) -> f32

Returns the global z order of this UiTransform as computed by the UiTransformSystem.