Struct amethyst_ui::UiTransformBuilder

pub struct UiTransformBuilder {
    pub id: String,
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub width: f32,
    pub height: f32,
    pub tab_order: i32,
    pub opaque: bool,
    pub percent: bool,
    pub stretch: Option<Stretch>,
    pub anchor: Anchor,
    pub mouse_reactive: bool,
    pub hidden: bool,
}

Loadable UiTransform data. By default z is equal to one.

Fields
id: String

An identifier. Serves no purpose other than to help you distinguish between UI elements.

x: f32

X coordinate

y: f32

Y coordinate

z: f32

Z coordinate, defaults to one.

width: f32

The width of this UI element.

height: f32

The height of this UI element.

tab_order: i32

The UI element tab order. When the player presses tab the UI focus will shift to the UI element with the next highest tab order, or if another element with the same tab_order as this one exists they are ordered according to Entity creation order. Shift-tab walks this ordering backwards.

opaque: bool

Indicates if actions on the ui can go through this element. If set to false, the element will behaves as if it was transparent and will let events go to the next element (for example, the text on a button).

percent: bool

Renders this UI element by evaluating transform as a percentage of the parent size, rather than rendering it with pixel units.

stretch: Option<Stretch>

If a child ui element needs to fill its parent this can be used to stretch it to the appropriate size.

anchor: Anchor

Indicates where the element sits, relative to the parent (or to the screen, if there is no parent)

mouse_reactive: bool

Allow mouse events on this UI element.

hidden: bool

Hides an entity by adding a HiddenPropagate component

Methods

impl UiTransformBuilder

pub fn with_id<S>(self, id: S) -> Self where
S: ToString,

Set id

pub fn with_position(self, x: f32, y: f32, z: f32) -> Self

Set position

pub fn with_size(self, width: f32, height: f32) -> Self

Set size

pub fn with_tab_order(self, tab_order: i32) -> Self

Set tab order

pub fn transparent(self) -> Self

Set to event transparent

pub fn hide(self) -> Self

Hides an entity by adding a HiddenPropagate component

pub fn reactive(self) -> Self

Add mouse reactive

pub fn with_anchor(self, anchor: Anchor) -> Self

Set anchor

pub fn with_stretch(self, stretch: Stretch) -> Self

Set stretch