Crate amethyst_ui

Provides components and systems to create an in game user interface.

Structs

DrawUi
Draw Ui elements. UI won't display without this. It's recommended this be your last pass.

FontAsset
A loaded set of fonts from a file.

ResizeSystem
This system rearranges UI elements whenever the screen is resized using their UiResize component.

TextEditing
If this component is attached to an entity with a UiText then that UiText is editable. This component also controls how that editing works.

TtfFormat
Loads font files, supports TrueType and some OpenType files.

UiBundle
UI bundle

UiFocused
This resource stores the currently focused UI element.

UiImage
A component with the texture to display in this entity's UiTransform

UiResize
Whenever the window is resized the function in this component will be called on this entity's UiTransform, along with the new width and height of the window.

UiSystem
This system processes the underlying UI data as needed.

UiText
A component used to display text in this entity's UiTransform

UiTransform
The raw pixels on screen that are populated.

Type Definitions

FontHandle
A handle to font data stored with amethyst_assets.

OtfFormat
Identical to TtfFormat.